faceRecSys/requirements.txt
matplotlib==3.2.2
numpy==1.19.5
opencv_python==4.5.2.52
Pillow==9.1.0
torch==1.9.0
torchvision==0.10.0
